Analytic as a service/GeosysPy-develop_RADAR_PPUS/requirements-dev.txt
python-dotenv
pytest
flake8
black
twine
mkdocs
mkdocstrings